Reorganize URL parameter structs
Summary:
The main module of the API server is growing pretty large, since a new method and query parameters struct is required for each new HTTP endpoint added to the server. To make this manageable, let's restructure the code so that each handler function is paired with a params struct that contains its corresponding params.

This makes it easier to make changes to individual handlers, such as switching the method GET to a POST and changing the handler to read the params from the request body.

Reviewed By: StanislavGlebik

Differential Revision: D14222620

fbshipit-source-id: 67abea2ff32a3d6cc278375efd368fa2969ab891