mononoke: fix combinatorial explosion when creating commits for mergy repos
Summary:
Some repos that are not that  big but they have a lot of merges
And sometimes there are a lot of merges where one merge parent
is ancestor of another merge parent, see example below:

  o  <- A
  |\
  | o <- B
  |/|
  o | <- C
  |\|
  | o
  |/|
  o |
  |\|
  ...
  |/|
  | ~
  o
  |\
  ~ ~

Turned out in the cases like that blobimport or push can hit a combinatorial explosion.
See an example. Commit A depends on commit B and C. So a future that creates a commit A also
need to wait until commit B and commit C are created i.e. roughly future for commit A looks like
`B.join(C).and_then(...)`. So when future A is polled it polls future B and after it polls future C.
The problem is that future B also polls future C and they do so all the way
down. So we poll every future a lot of times even when it's completely unnecessary.
This is similar to doing DFS without remembering visited nodes. In that case
event loop spends all the time polling nodes over and over again.

Note that combinatorial explosion depends and how many in-flight futures we
have. For example, reducing the number of commits that are computed simultaneously
(see `CONCURRENT_CHANGESETS` parameter of blobimport) avoids the issue, but
obviously it slows down blobimport for "normal" non-mergy repos. It's also
harder to repro in debug mode.

Differential Revision: D14100259

fbshipit-source-id: 8bcbc34a2843477f3704eb43cf5876e16ad4be61