move methods from fb303.FacebookService to fb303_core.BaseService
Summary:
Move most of the fb303 methods from `fb303.FacebookService` to
`fb303_core.BaseService`.

A few methods have been left behind.  In most cases these are methods that are
either already deprecated or that we would like to deprecate in the future.

One important method that is not being moved as part of this diff is
`getStatus()`.  We do want to move this to `BaseService` in the future, but
doing so is a bit more challenging than the other methods, since it also uses
a custom enum that will need to be moved from `fb303.thrift` to
`fb303_core.thrift`.  In most languages this can be done relatively easily
by replacing it with a typedef in `fb303.thrift`.  Unfortunately the `py3`
thrift generator does not handle typedefs well, so more work will need to be
done to enable this for python3 thrift code.

Reviewed By: yfeldblum

Differential Revision: D13688965

fbshipit-source-id: 48d3d007ef6b63b4176cab753624813b7f9d6526