Fix conversion between HgHistoryEntry and LooseHistoryEntry
Summary:
When converting from Mononoke `HgHistoryEntry` to Mercurial's `LooseHistoryEntry`, rather than directly copying the parents, we need to implement remotefilelog's logic for handling copied files.

Specifically, a copied file will usually have 0 parents (or 1 in the case of a merge). The filenode of the original (pre-copy) file is stored separately in the file's copy info. In the remtoefilelog loose file format, the filenode of the original file is stored as the p1 for the copy file, and the merge parent (if any) of the file is stored as the p2.

Reviewed By: StanislavGlebik

Differential Revision: D14256997

fbshipit-source-id: 09a070fb95d8b6c08c26e73156123d33cbfc685d