strip mutation extras when pushrebasing commits
Summary:
In the future we will be introducing mutation recording in commit extras.  For
pushrebase, these fields need to be stripped when the commit is rebased.  This
matches the Mercurial server behaviour.

At some point we will update Mononoke pushrebase to fill in these fields with
the pushrebase information, but that can wait until it is more widely deployed.

Reviewed By: quark-zju

Differential Revision: D14263856

fbshipit-source-id: 78b682fc580695137fd1fd4a4f65210a77f16b24