update "eden doctor" to report out-of-date working directories
Summary:
Make "eden doctor" check if the process's current working directory matches
the "$PWD" environment variable, and suggest that the user run "cd / && cd -"
if they do not match.

The most common cause of this is users that cd'ed into an Eden checkout
directory before Eden was running, and then started Eden but haven't updated
their working directory since.  In this shell they will still see the contents
of the underlying mount point directory itself and not the Eden checkout
contents that have been mounted on top of it.

Reviewed By: chadaustin

Differential Revision: D14249909

fbshipit-source-id: ecfc2a56180e3a6161940e9e1ec6d94663bddbc0