hgcli: pass ssh related env variables to Mononoke
Summary:
This additional data that is send to Mononoke will be used e.g. in verify_integrity hook.
This diff also contains change to the user_unixname, instead of passing env variable USER there we pass value produced by users crate. Since the user_unixname is not used anywhere now it is safe to do so.

Reviewed By: StanislavGlebik

Differential Revision: D14385280

fbshipit-source-id: 1e48c232fafbba5d5188c7d162e9ef21efd738f7