mononoke/glusterblob: use a shallower directory tree (FORMAT BREAKING)
Summary:
Reduce directory fan out levels to 2 wide ones (xxx/xxx) rather than 4 narrow
(xx/xx/xx/xx). This saves write latency cost: with 4 levels, until we get
~billions of blobs, every write will also be a mkdir, and we'll have many
single-entry directories. With 2 wide levels this will be true up to ~millions of
blobs, but then start to fill out the leaf directories.

The downside is that the leaf directory loading will be higher once we do have
billions of objects, but I think it will be managable.

This is NOT INTEROPERABLE with 4-level fanout, so invalidates existing stores.

Reviewed By: aslpavel

Differential Revision: D13923991

fbshipit-source-id: b93ddc49305c921f8a1b34606796b6d75273fb75